test: colocate tests with sources
