+++ src/core/BUILD.bazel
@@ -1,4 +1,4 @@
-load("//:ts.bzl", "ts_library")
+load("//:ts.bzl", "ts_library", "ts_test")
 
 ts_library(
     name = "core",
@@ -15,5 +15,16 @@ ts_library(
         "//src/internal",
         "//src/internal/decoders",
         "//src/internal/utils",
+    ],
+)
+
+ts_test(
+    name = "core_test",
+    srcs = ["streaming.test.ts"],
+    tsconfig_types = ["node"],
+    deps = [
+        ":core",
+        "//:node_modules/@types/node",
+        "//src/internal",
     ],
 )
