+++ src/internal/BUILD.bazel
@@ -1,4 +1,4 @@
-load("//:ts.bzl", "ts_library")
+load("//:ts.bzl", "ts_library", "ts_test")
 
 ts_library(
     name = "internal",
@@ -22,5 +22,20 @@ ts_library(
         "//src",
         "//src/core",
         "//src/internal/utils",
+    ],
+)
+
+ts_test(
+    name = "internal_test",
+    srcs = [
+        "headers.test.ts",
+        "to-file.test.ts",
+        "uploads.test.ts",
+    ],
+    tsconfig_types = ["node"],
+    deps = [
+        ":internal",
+        "//:node_modules/@types/node",
+        "//src/core",
     ],
 )
