+++ src/internal/decoders/BUILD.bazel
@@ -1,8 +1,14 @@
-load("//:ts.bzl", "ts_library")
+load("//:ts.bzl", "ts_library", "ts_test")
 
 ts_library(
     name = "decoders",
     srcs = ["line.ts"],
     visibility = ["//visibility:public"],
     deps = ["//src/internal/utils"],
+)
+
+ts_test(
+    name = "decoders_test",
+    srcs = ["line.test.ts"],
+    deps = [":decoders"],
 )
