+++ src/internal/headers.test.ts
@@ -1,9 +1,5 @@
 import { inspect } from 'node:util';
-import {
-  buildHeaders,
-  type HeadersLike,
-  type NullableHeaders,
-} from '@perplexity-ai/perplexity_ai/internal/headers';
+import { buildHeaders, type HeadersLike, type NullableHeaders } from './headers.js';
 
 function inspectNullableHeaders(headers: NullableHeaders) {
   return `NullableHeaders {${[
