+++ src/internal/uploads.test.ts
@@ -1,5 +1,5 @@
-import { multipartFormRequestOptions, createForm } from '@perplexity-ai/perplexity_ai/internal/uploads';
-import { toFile } from '@perplexity-ai/perplexity_ai/core/uploads';
+import { multipartFormRequestOptions, createForm } from './uploads.js';
+import { toFile } from '../core/uploads.js';
 
 describe('form data validation', () => {
   test('valid values do not error', async () => {
