+++ src/resources/BUILD.bazel
@@ -1,4 +1,4 @@
-load("//:ts.bzl", "ts_library")
+load("//:ts.bzl", "ts_library", "ts_test")
 
 ts_library(
     name = "resources",
@@ -21,5 +21,18 @@ ts_library(
         "//src/resources/browser",
         "//src/resources/chat",
         "//src/resources/responses",
+    ],
+)
+
+ts_test(
+    name = "resources_test",
+    srcs = [
+        "contextualized-embeddings.test.ts",
+        "embeddings.test.ts",
+        "search.test.ts",
+    ],
+    deps = [
+        ":resources",
+        "//src",
     ],
 )
