+++ src/resources/browser/BUILD.bazel
@@ -1,4 +1,4 @@
-load("//:ts.bzl", "ts_library")
+load("//:ts.bzl", "ts_library", "ts_test")
 
 ts_library(
     name = "browser",
@@ -13,5 +13,14 @@ ts_library(
         "//src/internal",
         "//src/internal/utils",
         "//src/resources",
+    ],
+)
+
+ts_test(
+    name = "browser_test",
+    srcs = ["sessions.test.ts"],
+    deps = [
+        ":browser",
+        "//src",
     ],
 )
