+++ src/resources/chat/BUILD.bazel
@@ -1,4 +1,4 @@
-load("//:ts.bzl", "ts_library")
+load("//:ts.bzl", "ts_library", "ts_test")
 
 ts_library(
     name = "chat",
@@ -12,5 +12,14 @@ ts_library(
         "//src/core",
         "//src/internal",
         "//src/resources",
+    ],
+)
+
+ts_test(
+    name = "chat_test",
+    srcs = ["completions.test.ts"],
+    deps = [
+        ":chat",
+        "//src",
     ],
 )
