+++ src/resources/responses/BUILD.bazel
@@ -1,4 +1,4 @@
-load("//:ts.bzl", "ts_library")
+load("//:ts.bzl", "ts_library", "ts_test")
 
 ts_library(
     name = "responses",
@@ -14,5 +14,17 @@ ts_library(
         "//src/internal/utils",
         "//src/lib",
         "//src/resources",
+    ],
+)
+
+ts_test(
+    name = "responses_test",
+    srcs = [
+        "files.test.ts",
+        "responses.test.ts",
+    ],
+    deps = [
+        ":responses",
+        "//src",
     ],
 )
