--- tests/api-resources/BUILD.bazel
@@ -1,10 +1,0 @@
-load("//:ts.bzl", "ts_test")
-
-ts_test(
-    name = "api-resources_test",
-    srcs = [
-        "contextualized-embeddings.test.ts",
-        "embeddings.test.ts",
-        "search.test.ts",
-    ],
-)
