--- tests/api-resources/browser/BUILD.bazel
@@ -1,6 +1,0 @@
-load("//:ts.bzl", "ts_test")
-
-ts_test(
-    name = "browser_test",
-    srcs = ["sessions.test.ts"],
-)
