--- tests/api-resources/responses/BUILD.bazel
@@ -1,9 +1,0 @@
-load("//:ts.bzl", "ts_test")
-
-ts_test(
-    name = "responses_test",
-    srcs = [
-        "files.test.ts",
-        "responses.test.ts",
-    ],
-)
