--- tests/internal/decoders/BUILD.bazel
@@ -1,6 +1,0 @@
-load("//:ts.bzl", "ts_test")
-
-ts_test(
-    name = "decoders_test",
-    srcs = ["line.test.ts"],
-)
